+++ kotlin/reakt-native-toolkit/src/iosMain/kotlin/de/voize/reaktnativetoolkit/util/PromiseUtil.kt
@@ -13,7 +13,7 @@ inline fun <reified T> PromiseIOS.runCatchingWithArguments(action: () -> T, argu
         .runCatching { argumentsTransform(action()) }
         .onSuccess(resolve)
         .onFailure {
-            reject(it, exceptionInterceptor(it))
+            reject(it, it.getJSExtraData())
         }
 }